add: ServiceLocator
ServiceLocator is an anti-pattern.
But sometimes it is the powerful pattern to resolve another tricky
code design.
One of the useful scenario is that, there are many different types in
StorageItemView (SequenceStorageItemView, AudioStorageItemView, ...) and
those Views behavior differently.
When double-clicking SequenceStorageItemView creates SequenceView
if TimelineWidget exists. This needs WidgetsService to check whether
TimelineWidget is exists or not.
For AudioStorageItemView, it can perform action replacing the resource
with another audio resource, and this needs ImportService.
Since these views needs many different services on demand,
passing ServiceLocator to ViewFactory enables us to pick necessary
services from the ServiceLocator.